requirements.txt
netCDF4==1.4.1
coloredlogs==6.1
progressbar2==3.50.1
